сделал дейкстру универсальным

+++ GraphEditor/Model/WorkWithFile/Reader.cs
@@ -67,7 +67,7 @@ namespace Model.WorkWithFile
             while (line != null)
             {
                 string[] nodesLength = line.Split(";");
-                coordinats.Add(nodesLength[0], new Point(double.Parse(nodesLength[1]), double.Parse(nodesLength[2])));
+                coordinats.Add(nodesLength[0], new Point(int.Parse(nodesLength[1]), int.Parse(nodesLength[2])));
                 line = streamReader.ReadLine();
             }
